policy accounts transactions on activate

+++ accounts/transactions/transactions.php
@@ -2,12 +2,12 @@
 include("../../include/main.php");
 include("../../include/tables.php");
 
-$db = new Main(1,'UTF-8');
+$db = new Main(1, 'UTF-8');
 $db->admin_title = "Transactions";
 
 $db->show_header();
 
-$table = new draw_table('ac_transactions','actrn_transaction_ID','ASC');
+$table = new draw_table('ac_transactions', 'actrn_transaction_ID', 'ASC');
 
 $table->generate_data();
 
@@ -23,9 +23,9 @@ $table->generate_data();
                 <table class="table table-hover">
                     <thead>
                     <tr>
-                        <th scope="col"><?php $table->display_order_links('ID','actrn_transaction_ID');?></th>
-                        <th scope="col"><?php $table->display_order_links('Type','actrn_type');?></th>
-                        <th scope="col"><?php $table->display_order_links('Code','actrn_status');?></th>
+                        <th scope="col"><?php $table->display_order_links('ID', 'actrn_transaction_ID'); ?></th>
+                        <th scope="col"><?php $table->display_order_links('Type', 'actrn_type'); ?></th>
+                        <th scope="col"><?php $table->display_order_links('Code', 'actrn_status'); ?></th>
                         <th scope="col">
                             <a href="transaction_modify.php">
                                 <i class="fas fa-plus-circle"></i>
@@ -38,19 +38,40 @@ $table->generate_data();
                     while ($row = $table->fetch_data()) {
                         ?>
                         <tr>
-                            <th scope="row"><?php echo $row["actrn_transaction_ID"];?></th>
-                            <td><?php echo $row["actrn_type"];?></td>
-                            <td><?php echo $row["actrn_status"];?></td>
+                            <th scope="row"><?php echo $row["actrn_transaction_ID"]; ?></th>
+                            <td><?php echo $row["actrn_type"]; ?></td>
+                            <td><?php echo $row["actrn_status"]; ?></td>
                             <td>
-                                <a href="transaction_modify.php?lid=<?php echo $row["actrn_transaction_ID"];?>"><i class="fas fa-edit"></i></a>&nbsp
+                                <?php
+                                if ($row['actrn_status'] == 'Outstanding') {
+                                    ?>
+                                    <a href="transaction_modify.php?lid=<?php echo $row["actrn_transaction_ID"]; ?>"><i
+                                                class="fas fa-edit"></i></a>&nbsp
+                                <?php } ?>
 
-                                <a href="transaction_delete.php?lid=<?php echo $row["actrn_transaction_ID"];?>"
-                                   onclick="return confirm('Are you sure you want to delete this Transaction?');"><i class="fas fa-minus-circle"></i></a>
+                                <?php
+                                if ($row['actrn_status'] == 'Locked' || $row['actrn_status'] == 'Active') {
+                                    ?>
+                                    <a href="transaction_modify.php?lid=<?php echo $row["actrn_transaction_ID"]; ?>"><i
+                                                class="fas fa-eye"></i></a>&nbsp
+                                <?php } ?>
 
-                                <a href="transaction_change_status.php?lid=<?php echo $row['actrn_transaction_ID'];?>">
-                                    <i class="fas fa-lock"
-                                       title="<?php echo $db->showLangText('Activate Transaction','Ενεργοποίηση Πράξης');?>"></i>
-                                </a>
+                                <?php
+                                if ($row['actrn_status'] == 'Outstanding') {
+                                    ?>
+                                    <a href="transaction_delete.php?lid=<?php echo $row["actrn_transaction_ID"]; ?>"
+                                       onclick="return confirm('Are you sure you want to delete this Transaction?');"><i
+                                                class="fas fa-minus-circle"></i></a>
+                                <?php } ?>
+
+                                <?php
+                                if ($row['actrn_status'] == 'Outstanding' || $row['actrn_status'] == 'Locked') {
+                                    ?>
+                                    <a href="transaction_change_status.php?lid=<?php echo $row['actrn_transaction_ID']; ?>">
+                                        <i class="fas fa-lock"
+                                           title="<?php echo $db->showLangText('Activate Transaction', 'Ενεργοποίηση Πράξης'); ?>"></i>
+                                    </a>
+                                <?php } ?>
 
                             </td>
                         </tr>
